# encoding/mvt: use protoscan for unmarshalling