--- a/encoding/mvt/marshal.go
+++ b/encoding/mvt/marshal.go
@@ -4,7 +4,6 @@ import (
 	"bytes"
 	"compress/gzip"
 	"fmt"
-	"io/ioutil"
 	"strconv"
 
 	"github.com/paulmach/orb/encoding/mvt/vectortile"
@@ -84,72 +83,6 @@ func Marshal(layers Layers) ([]byte, error) {
 	return proto.Marshal(vt)
 }
 
-// UnmarshalGzipped takes gzipped Mapbox Vector Tile (MVT) data and unzips it
-// before decoding it into a set of layers, It does not project the coordinates.
-func UnmarshalGzipped(data []byte) (Layers, error) {
-	gzreader, err := gzip.NewReader(bytes.NewBuffer(data))
-	if err != nil {
-		return nil, errors.WithMessage(err, "failed to create gzreader")
-	}
-
-	decoded, err := ioutil.ReadAll(gzreader)
-	if err != nil {
-		return nil, errors.WithMessage(err, "failed to unzip")
-	}
-
-	return Unmarshal(decoded)
-}
-
-// Unmarshal takes Mapbox Vector Tile (MVT) data and converts into a
-// set of layers, It does not project the coordinates.
-func Unmarshal(data []byte) (Layers, error) {
-	vt := &vectortile.Tile{}
-	err := vt.Unmarshal(data)
-	if err != nil {
-		return nil, err
-	}
-
-	return decode(vt)
-}
-
-func decode(vt *vectortile.Tile) (Layers, error) {
-	result := make(Layers, 0, len(vt.Layers))
-	for i, l := range vt.Layers {
-		layer := &Layer{
-			Name:     l.GetName(),
-			Version:  l.GetVersion(),
-			Extent:   l.GetExtent(),
-			Features: make([]*geojson.Feature, 0, len(l.Features)),
-		}
-
-		for j, f := range l.Features {
-			geom, err := decodeGeometry(f.GetType(), f.Geometry)
-			if err != nil {
-				return nil, errors.WithMessage(err, fmt.Sprintf("layer %d: feature %d", i, j))
-			}
-
-			properties := decodeFeatureProperties(l.Keys, l.Values, f.Tags)
-
-			if geom != nil {
-				gjf := &geojson.Feature{
-					Geometry:   geom,
-					Properties: properties,
-				}
-
-				if f.Id != nil {
-					gjf.ID = float64(*f.Id)
-				}
-
-				layer.Features = append(layer.Features, gjf)
-			}
-		}
-
-		result = append(result, layer)
-	}
-
-	return result, nil
-}
-
 func encodeProperties(kve *keyValueEncoder, properties geojson.Properties) ([]uint32, error) {
 	tags := make([]uint32, 0, 2*len(properties))
 	for k, v := range properties {
@@ -163,36 +96,6 @@ func encodeProperties(kve *keyValueEncoder, properties geojson.Properties) ([]ui
 	}
 
 	return tags, nil
-}
-
-func decodeFeatureProperties(
-	keys []string,
-	values []*vectortile.Tile_Value,
-	tags []uint32,
-) geojson.Properties {
-	result := make(geojson.Properties, len(tags)/2)
-	if len(tags) == 0 {
-		return result
-	}
-
-	for i := 2; i <= len(tags); i += 2 {
-		vi := tags[i-1]
-		if int(vi) >= len(values) {
-			continue
-		}
-
-		v := decodeValue(values[vi])
-		if v != nil {
-			ti := tags[i-2]
-			if int(ti) >= len(keys) {
-				continue
-			}
-
-			result[keys[ti]] = v
-		}
-	}
-
-	return result
 }
 
 func convertID(id interface{}) *uint64 {