--- a/go.mod
+++ b/go.mod
@@ -4,5 +4,6 @@ go 1.13
 
 require (
 	github.com/gogo/protobuf v1.3.1
+	github.com/paulmach/protoscan v0.1.0
 	github.com/pkg/errors v0.9.1
 )