… encoding/mvt/geometry.go
… go.sum