Vector Tiles have an [encoding](https://github.com/mapbox/vector-tile-spec/tree/master/2.1#43-geometry-encoding) for the geometry. This causes an intermediate state were the protobuf is unpacked into a `[]uint32` and then converted into the `orb.Geometry` types. This intermediate step can be removed by using a low lever decoder like [protoscan](https://github.com/paulmach/protoscan)

The behavior is the same, but the performance is substantially better:
```
benchmark                     old ns/op     new ns/op     delta
BenchmarkUnmarshal-12         367590        246198        -33.02%

benchmark                     old allocs     new allocs     delta
BenchmarkUnmarshal-12         6629           2476           -62.65%

benchmark                     old bytes     new bytes     delta
BenchmarkUnmarshal-12         372954        211681        -43.24%
```